+++ mry/tests/integration/main.rs
@@ -17,7 +17,8 @@ mod partial_mock;
 mod reference_and_pattern;
 mod returns_with_recursive_call;
 mod simple_case;
-mod skip;
+mod skip_arg;
+mod skip_methods;
 mod static_function;
 
 #[cfg(feature = "send_wrapper")]